error strim development

+++ gemdaq-testing/gemsupervisor/include/gem/supervisor/GEMGLIBSupervisorWeb.h
@@ -233,7 +233,7 @@ namespace gem {
         gem::readout::GEMDataParker* gemDataParker;
 
         // Counter
-        int counter_[3];
+        uint64_t counter_[3];
 
         // VFAT Blocks Counter
         int vfat_;
